% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdog.params.r
\name{asdog.info.params}
\alias{asdog.info.params}
\title{get asdog params info}
\usage{
asdog.info.params(...)
}
\arguments{
\item{...}{names of parameters (if empty return info for all parameters)}
}
\description{
get asdog params info
}

