% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdog.main.r
\name{asdog.main.params}
\alias{asdog.main.params}
\title{get asdog parameters from args and (optionally)
from previous params}
\usage{
asdog.main.params(args, params = NULL)
}
\description{
get asdog parameters from args and (optionally)
from previous params
}

