% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdog.params.r
\name{asdog.new.params}
\alias{asdog.new.params}
\title{new asdog.model parameters with defaults}
\usage{
asdog.new.params(..., as.args = FALSE, .default = asdog.default.params)
}
\description{
new asdog.model parameters with defaults
}
\details{
you may specify new default with ... args : key=value

to import default from another params object :
  do.call(asdog.new.params, params)

to change program defaults : .default=<function>
}
\note{
this function is usually called from \link{asdog.main.params}
so there is no need to call it directly
}
\seealso{
\link{asdog.main.params}
}

