% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdog.gccorrect.report.r
\name{asdog.tex.report.gccorrect}
\alias{asdog.tex.report.gccorrect}
\title{produce gccorrect tex report}
\usage{
asdog.tex.report.gccorrect(tex, correct)
}
\description{
produce gccorrect tex report
}

