% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdog.gcmodel.report.r
\name{asdog.tex.report.gcmodel}
\alias{asdog.tex.report.gcmodel}
\title{produce gccor.model tex report}
\usage{
asdog.tex.report.gcmodel(tex, model)
}
\description{
produce gccor.model tex report
}

