% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdog.plmodel.report.r
\name{asdog.tex.report.plmodel}
\alias{asdog.tex.report.plmodel}
\title{produce plody model tex report}
\usage{
asdog.tex.report.plmodel(tex, model)
}
\description{
produce plody model tex report
}

